+++ src/templates/BlogPost.js
@@ -59,7 +59,7 @@ const BlogPostTemplate = ({ data }) => {
           />
         </article>
 
-        <div className="relative max-w-lg mx-auto lg:max-w-2xl">
+        <div className="relative max-w-lg mx-auto lg:max-w-xl">
           <SubscribeToNewsletterCTA
             setModalOpen={setModalOpen}
             email={email}
